+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/amplify-meta-updaters.ts
@@ -106,18 +106,28 @@ export const getPostUpdateAuthMetaUpdater = (context: any) => async (resourceNam
   return resourceName;
 };
 
-function getFrontendConfig(authParameters: AuthParameters) {
-  const loginMechanisms = (authParameters?.aliasAttributes || []).map((att: string) => att.toUpperCase());
+export function getFrontendConfig(authParameters: AuthParameters) {
   const verificationMechanisms = (authParameters?.autoVerifiedAttributes || []).map((att: string) => att.toUpperCase());
+  const loginMechanisms = new Set<string>();
+  (authParameters?.aliasAttributes ?? []).forEach(it => loginMechanisms.add(it.toUpperCase()));
+
+  // backwards compatibility
+  if (authParameters?.usernameAttributes && authParameters.usernameAttributes.length > 0) {
+    authParameters.usernameAttributes[0].split(',').forEach(it => loginMechanisms.add(it.trim().toUpperCase()));
+  }
 
   if (authParameters.authProviders) {
     authParameters.authProviders.forEach((provider: string) => {
       let name = authProviderList.find(it => it.value === provider)?.name;
 
       if (name) {
-        loginMechanisms.push(name.toUpperCase());
+        loginMechanisms.add(name.toUpperCase());
       }
     });
+  }
+
+  if (loginMechanisms.size === 0) {
+    loginMechanisms.add('PREFERRED_USERNAME');
   }
 
   const signupAttributes = (authParameters?.requiredAttributes || []).map((att: string) => att.toUpperCase());
@@ -139,7 +149,7 @@ function getFrontendConfig(authParameters: AuthParameters) {
   }
 
   return {
-    loginMechanisms: loginMechanisms,
+    loginMechanisms: Array.from(loginMechanisms),
     signupAttributes: signupAttributes,
     passwordProtectionSettings: passwordProtectionSettings,
     mfaConfiguration: authParameters?.mfaConfiguration,