+++ packages/amplify-graphql-transformer-interfaces/src/transformer-context/resource-resource-provider.ts
@@ -1,4 +1,5 @@
 export interface TransformerResourceHelperProvider {
   generateTableName(modelName: string): string;
+  generateRoleName(baseName: string): string;
   registerModelToTableNameMaping(modelName: string, tableName: string): void;
 }