+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.21](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.20...amplify-category-analytics@2.21.21) (2021-09-18)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 ## [2.21.20](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.19...amplify-category-analytics@2.21.20) (2021-09-14)
 
 **Note:** Version bump only for package amplify-category-analytics