+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.21](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.20...amplify-nodejs-function-template-provider@1.6.21) (2021-09-18)
+
+**Note:** Version bump only for package amplify-nodejs-function-template-provider
+
+
+
+
+
 ## [1.6.20](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.19...amplify-nodejs-function-template-provider@1.6.20) (2021-09-14)
 
 **Note:** Version bump only for package amplify-nodejs-function-template-provider